RSS scaffolding and test
RSS feed is fetching! As of right now the files for the project just contain the scaffolding with the test of actually fetching the RSS feed being in scratch.py. All I need to do now is translate the code from scratch.py into functions in the class in rss.py and then use that to fill the database! I can see timezones being an issue later...