Search bar!
Tagging, commenting (posting and deleting), comment voting, and searching by either keyword in the blurb title/description or by a tag is complete! Now to just move these additions onto the home page (not the search bar, that's just for the discover page) so that they aren't just on the discovery page, and then cleaning up CSS- then there's just fixing the profile image bug and then RSS is back on and we're good to deploy!